+++ simple-cms/astro/.env.example
@@ -3,3 +3,4 @@ DIRECTUS_PUBLIC_TOKEN=your-public-token       # Token for public content
 DIRECTUS_FORM_TOKEN=your-form-token           # Token for forms
 NEXT_PUBLIC_SITE_URL=http://localhost:3000    # Application URL
 DRAFT_MODE_SECRET=your-draft-mode-secret      # Secret for preview mode
+NEXT_PUBLIC_ENABLE_VISUAL_EDITING=true # Enable visual editing
