+++ simple-cms/astro/src/components/blocks/Button.tsx
@@ -3,7 +3,7 @@ import { Button as ShadcnButton, buttonVariants } from '@/components/ui/button';
 import { type LucideIcon, ArrowRight, Plus } from 'lucide-react';
 import { cn } from '@/lib/utils';
 
-import { setAttr } from '@/lib/directus/visual-editing-helper';
+import { setAttr } from '@directus/visual-editing';
 
 export interface ButtonProps {
   id: string;
