+++ simple-cms/astro/src/components/blocks/RichText.tsx
@@ -2,7 +2,8 @@ import React from 'react';
 import Tagline from '../ui/Tagline';
 import Headline from '@/components/ui/Headline';
 import Text from '@/components/ui/Text';
-import { setAttr, useDirectusVisualEditing } from '@/lib/directus/visual-editing-helper';
+import { setAttr } from '@directus/visual-editing';
+import { useDirectusVisualEditing } from '@/lib/directus/useDirectusVisualEditing';
 
 interface RichTextProps {
   data: {
