+++ simple-cms/astro/src/components/forms/DynamicForm.tsx
@@ -6,7 +6,7 @@ import Field from './FormField';
 import { buildZodSchema } from '@/lib/zodSchemaBuilder';
 import type { FormField as FormFieldType } from '@/types/directus-schema';
 import React from 'react';
-import { setAttr } from '@/lib/directus/visual-editing-helper';
+import { setAttr } from '@directus/visual-editing';
 
 interface DynamicFormProps {
   fields: FormFieldType[];
